Moved File Reader test classes to appropriate directories
