+++ src/test/java/ntnu/idatt2003/group27/utils/filehandler/csv/PlayerCsvFileReaderTest.java
@@ -1,4 +1,4 @@
-package ntnu.idatt2003.group27.models;
+package ntnu.idatt2003.group27.utils.filehandler.csv;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -11,12 +11,12 @@ import java.io.StringReader;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.IntStream;
 import javafx.scene.paint.Color;
+import ntnu.idatt2003.group27.models.Piece;
+import ntnu.idatt2003.group27.models.Player;
 import ntnu.idatt2003.group27.models.exceptions.ExceededMaxPlayersException;
 import ntnu.idatt2003.group27.models.exceptions.MissingPlayerException;
 import ntnu.idatt2003.group27.utils.RandomColor;
-import ntnu.idatt2003.group27.utils.filehandler.csv.PlayerCsvFileReader;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
